Finial Code Development

+++ SaveStarMoneyAPI/Dtos/PercentSavingDto.cs
@@ -3,7 +3,7 @@
     public class PercentSavingDto
     {
         public string SavingTitle { get; set; } = string.Empty;
-        public int Percentage { get; set; }
+        public decimal Percentage { get; set; }
         public string Frequency { get; set; } = string.Empty;
         public string Duration { get; set; } = string.Empty;
     }
