+++ SaveStarMoneyAPI/Entities/PercentageSaving.cs
@@ -5,7 +5,7 @@
         public long Id { get; set; }
         public string Name { get; set; } = string.Empty;
         public string SavingTitle { get; set; } = string.Empty;
-        public int Percentage { get; set; }
+        public decimal Percentage { get; set; }
         public string Frequency { get; set; } = string.Empty;
         public string Duration { get; set; } = string.Empty;
         public decimal PercentageAmount { get; set; }
